Move logic to determine if someone is following someone out of MainActivity
